kbenjell

+++ ft_calloc.c
@@ -16,7 +16,7 @@ void	*ft_calloc(size_t count, size_t size)
 {
 	void	*ptr;
 	char	*fill;
-	int		counter;
+	size_t		counter;
 
 	ptr = malloc(count * size);
 	if (count == 0 || size == 0)
